+++ python/wcs_utils/bin/log_example.py
@@ -1,0 +1,33 @@
+"""
+Copyright (c) Min.Wu - All Rights Reserved
+Unauthorized copying of this file, via any medium is strictly prohibited
+Proprietary and confidential
+Author: Min.Wu <wumin@126.com>, 2026/01/07
+"""
+
+
+import os
+import sys
+import time
+
+WCS_PY_ROOT = os.path.join(os.path.dirname(os.path.dirname(__file__)), "..")
+if os.path.exists(WCS_PY_ROOT):
+    sys.path.insert(0, WCS_PY_ROOT)
+
+
+from wcs_utils.logger.spdlog import logger, set_log_save_path  # noqa: E402
+
+
+def logger_example():
+    home_path = os.path.expanduser("~")
+    set_log_save_path(os.path.join(home_path, "wcs_logs/wcs_utils"))
+
+    logger.info("This is an info log message.")
+    logger.info("This is a formatted info log message: {}".format("Hello, World!"))
+    logger.info("This is a formatted info log message with args: {} and {}".format("test", 42))
+    logger.warning("This is a warning log message.")
+    logger.error("This is an error log message.")
+
+
+if __name__ == "__main__":
+    logger_example()