+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 2. #6 spdlog 测试
 3. #8 格式化
 4. #10 spdlog 封装成动态库
+5. #12 python 版本的spdlog